게시판 목록 수정 (완)

+++ FeatherWorld/src/main/resources/static/js/board/boardList.js
@@ -268,11 +268,8 @@ const renderBoardTypeList = async () => {
 
   boardTypeSidebar.append(div);
 
-  boardTypeList.forEach((boardType) => {
-    if (
-      boardType.authority == 0 ||
-      (boardType.authority == 1 && loginMemberNo === memberNo)
-    ) {
+  boardTypeList.forEach(boardType => {
+    if (boardType.authority == 0 || (boardType.authority == 1 && loginMemberNo === memberNo)) {
       const boardTypeItem = document.createElement("div");
       boardTypeItem.classList.add("board-type-item");
       boardTypeItem.dataset.boardCode = boardType.boardCode;
@@ -281,25 +278,32 @@ const renderBoardTypeList = async () => {
       boardTypeTitle.classList.add("board-type-title");
       boardTypeTitle.innerText = boardType.boardName;
 
-      const iconSpan = document.createElement("span");
+      // 기본 게시판에는 수정 허용 안됨!
+      if(boardType.boardCode !== defaultBoardCode) {
+        const iconSpan = document.createElement("span");
+        iconSpan.classList.add("icon-wrap");
 
-      const editIcon = document.createElement("span");
-      editIcon.classList.add("edit-icon");
-      const pencilIcon = document.createElement("span");
-      editIcon.classList.add("fa-solid", "fa-pencil");
+        const editIcon = document.createElement("span");
+        editIcon.classList.add("edit-icon");
+        const pencilIcon = document.createElement("span");
+        editIcon.classList.add("fa-solid", "fa-pencil");
 
-      editIcon.append(pencilIcon);
+        editIcon.append(pencilIcon);
 
-      const deleteIcon = document.createElement("span");
-      deleteIcon.classList.add("delete-icon");
-      const trashIcon = document.createElement("span");
-      trashIcon.classList.add("fa-solid", "fa-trash");
+        const deleteIcon = document.createElement("span");
+        deleteIcon.classList.add("delete-icon");
+        const trashIcon = document.createElement("span");
+        trashIcon.classList.add("fa-solid", "fa-trash");
 
-      deleteIcon.append(trashIcon);
+        deleteIcon.append(trashIcon);
 
-      iconSpan.append(editIcon, deleteIcon);
+        iconSpan.append(editIcon, deleteIcon);
 
-      boardTypeItem.append(boardTypeTitle, iconSpan);
+        boardTypeItem.append(boardTypeTitle, iconSpan);
+
+      } else {
+        boardTypeItem.append(boardTypeTitle);
+      }
 
       boardTypeSidebar.append(boardTypeItem);
     }
@@ -308,6 +312,10 @@ const renderBoardTypeList = async () => {
   document.querySelector(".board-type-sidebar").replaceWith(boardTypeSidebar);
 };
 
+/** 폴더 추가 버튼 생성
+ * @author Jiho
+ * @returns {HTMLDivElement}
+ */
 const createAddFolder = () => {
   const div = createDiv("add-folder", "Add Folder");
   const span = document.createElement("span");
@@ -426,7 +434,7 @@ if (leftSidebar) {
 
       const folderForm = {
         boardName: boardName,
-        authority: authority,
+        authority: authority
       };
 
       // 입력한 값들을 모두 ajax로 비동기 요청
@@ -446,6 +454,7 @@ if (leftSidebar) {
 
       // 현재 세션 갱신 후 화면에 게시판 목록 렌더링
       renderBoardTypeList().catch(console.error);
+      editMode = false;
 
       document.querySelector(".add-folder-form")?.remove();
       leftSidebar.append(createAddFolder());
@@ -460,7 +469,7 @@ if (leftSidebar) {
       editMode = !editMode;
 
       const iconSpan = document.querySelectorAll(
-        ".board-type-item > span:last-child"
+        ".board-type-item > .icon-wrap"
       );
 
       if(editMode) {
@@ -480,28 +489,76 @@ if (leftSidebar) {
     const editIcon = e.target.closest(".edit-icon");
     if(editMode && editIcon) {
 
-      // input으로 수정값 받아오기 필요
-      const boardName = "수정 이름";
-      const authority = 1;
+      const parentBoardItem = editIcon.parentElement.parentElement;
 
-      const folderForm = {
-        boardName: boardName,
-        authority: authority
-      };
+      const editBoardCode = parentBoardItem.dataset.boardCode;
 
+      // 기존 게시판명
+      const title = parentBoardItem.firstElementChild.innerText;
 
-      // 수정 모드 초기화
-      editMode = !editMode;
+      const input = document.createElement("input");
+      input.name = "boardName";
+      input.value = title;
+
+      async function updateBoardName() {
+        if (input.value.trim() === "") {
+          alert("게시판 이름을 입력하세요.");
+          return;
+        }
+
+        const folderForm = {
+          boardCode: editBoardCode,
+          boardName: input.value
+        };
+
+        const resp = await fetch(`/${memberNo}/board/update`, {
+          method: "put",
+          headers: { "Content-Type": "application/json" },
+          body: JSON.stringify(folderForm)
+        });
+        const result = await resp.text();
+
+        if (result == 0) {
+          alert("게시판 수정 실패");
+
+          renderBoardTypeList().catch(console.error);
+
+          // 수정 모드 초기화
+          editMode = false;
+          return;
+        }
+
+        alert("게시판을 수정했습니다.");
+        await renderBoardTypeList().catch(console.error);
+
+        // 수정 모드 초기화
+        editMode = false;
+      }
+
+      editIcon.addEventListener("click", updateBoardName);
+
+      parentBoardItem.firstElementChild.replaceWith(input);
+      input.focus();
+
+      return;
     }
 
     // 쓰레기통 아이콘 클릭시
     const deleteIcon = e.target.closest(".delete-icon");
     if(editMode && deleteIcon) {
 
+      const deleteBoardCode = deleteIcon.parentElement.parentElement.dataset.boardCode;
+
+      // 삭제되는 게시판이 기본 게시판인 경우
+      if(deleteBoardCode == defaultBoardCode) {
+        alert("기본 게시판은 삭제할 수 없습니다.");
+        return;
+      }
+
       const resp = await fetch(`/${memberNo}/board/delete`, {
         method: "delete",
         headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ boardCode: boardCode })
+        body: JSON.stringify({ boardCode: deleteBoardCode })
       });
       const result = await resp.text();
 
@@ -516,8 +573,6 @@ if (leftSidebar) {
 
       // 수정 모드 초기화
       editMode = !editMode;
-      // boardCode & 게시글 초기화 필요
-
     }
   });
 }
@@ -530,5 +585,5 @@ window.addEventListener("popstate", () => {
   loadBoardList(boardCode);
 });
 
-loadBoardList(boardCode);
+loadBoardList(defaultBoardCode);
 renderBoardTypeList().catch(console.error);
